refactor: move Hello sending to handshake module
